+++ src/const.js
@@ -1,0 +1,38 @@
+export const ages = [`0+`, `6+`, `12+`, `16+`, `18+`];
+
+export const allEmojies = [`angry`, `puke`, `sleeping`, `smile`];
+
+export const moviesTitles = [
+  `Made for each other.png`,
+  `Popeye meets sinbad.png`,
+  `Sagebrush trail.jpg`,
+  `Santa claus conquers the martians.jpg`,
+  `The dance of life.jpg`,
+  `The great flamarion.jpg`,
+  `The man with the golden arm.jpg`
+];
+
+export const genres = [
+  `Musical`,
+  `Western`,
+  `Drama`,
+  `Cartoon`,
+  `Mystery`,
+  `Comedy`,
+  `Horror`
+];
+
+export const descriptions = [
+  `In rutrum ac purus sit amet tempus`,
+  `Nunc fermentum tortor ac porta dapibus`,
+  `Aliquam erat volutpat`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui`,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante`,
+  `Aliquam id orci ut lectus varius viverra`,
+  `Fusce tristique felis at fermentum pharetra`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget`,
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit`
+];
+
