+++ src/mock/card.js
@@ -1,8 +1,11 @@
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
+import {getRandomInteger} from "../utils";
+import {ages, moviesTitles, genres, descriptions} from "../const";
 
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
+const generateRating = () => {
+  const integer = getRandomInteger(1, 10);
+  const fractional = getRandomInteger(0, 9);
+
+  return `${integer}.${fractional}`;
 };
 
 const generateDuration = () => {
@@ -13,45 +16,17 @@ const generateDuration = () => {
   return hours + ` ` + minutes;
 };
 
-const generateRating = () => {
-  const integer = getRandomInteger(1, 10);
-  const fractional = getRandomInteger(0, 9);
-
-  return `${integer}.${fractional}`;
-};
-
 const generateAgeLimits = () => {
-  const ages = [`0+`, `6+`, `12+`, `16+`, `18+`];
   const randomIndex = getRandomInteger(0, ages.length - 1);
   return ages[randomIndex];
-
 };
 
 const generateTitle = () => {
-  const titles = [
-    `Made for each other.png`,
-    `Popeye meets sinbad.png`,
-    `Sagebrush trail.jpg`,
-    `Santa claus conquers the martians.jpg`,
-    `The dance of life.jpg`,
-    `The great flamarion.jpg`,
-    `The man with the golden arm.jpg`
-  ];
-  const randomIndex = getRandomInteger(0, titles.length - 1);
-  return titles[randomIndex];
+  const randomIndex = getRandomInteger(0, moviesTitles.length - 1);
+  return moviesTitles[randomIndex];
 };
 
-const generateGenre = () => {
-  const genres = [
-    `Musical`,
-    `Western`,
-    `Drama`,
-    `Cartoon`,
-    `Mystery`,
-    `Comedy`,
-    `Horror`
-  ];
-
+const generateGenres = () => {
   const titles = new Set();
   for (let i = 0; i < getRandomInteger(1, 5); i++) {
     titles.add(genres[getRandomInteger(0, genres.length - 1)]);
@@ -61,20 +36,6 @@ const generateGenre = () => {
 };
 
 const generateDescription = () => {
-  const descriptions = [
-    `In rutrum ac purus sit amet tempus`,
-    `Nunc fermentum tortor ac porta dapibus`,
-    `Aliquam erat volutpat`,
-    `Sed sed nisi sed augue convallis suscipit in sed felis`,
-    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui`,
-    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum`,
-    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante`,
-    `Aliquam id orci ut lectus varius viverra`,
-    `Fusce tristique felis at fermentum pharetra`,
-    `Cras aliquet varius magna, non porta ligula feugiat eget`,
-    `Lorem ipsum dolor sit amet, consectetur adipiscing elit`
-  ];
-
   const sentences = new Set();
   for (let i = 0; i < getRandomInteger(1, 5); i++) {
     sentences.add(descriptions[getRandomInteger(0, descriptions.length - 1)]);
@@ -100,7 +61,7 @@ export const generateCard = () => {
     rating: generateRating(),
     releaseDate: generatedate(),
     duration: generateDuration(),
-    genre: generateGenre(),
+    genres: generateGenres(),
     description: generateDescription(),
     commentsSum: getRandomInteger(0, 5),
     watchPlan: Boolean(getRandomInteger(0, 1)),
